requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.2
groq==0.9.0
httpx==0.25.2
pydantic==2.5.0
python-dotenv==1.0.0
sse-starlette==1.8.2
prometheus-client==0.19.0